pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mkdocs-image-gallery-plugin"
version = "1.3.0"
description = "An MkDocs plugin to generate an image gallery from a folder of images"
readme = "README.md"
authors = [
    {name = "APinchofDill"}
]
license = "MIT"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
keywords = ["mkdocs", "image", "gallery"]
requires-python = ">=3.8"
dependencies = [
    "jinja2>=3.1.4",
    "mkdocs>=1.4.1",
    "PyYAML>=6.0.1"
]

[project.urls]
Homepage = "https://github.com/APinchofDill/mkdocs-image-gallery-plugin"

[project.entry-points."mkdocs.plugins"]
"image-gallery" = "mkdocs_image_gallery_plugin.plugin:ImageGalleryPlugin"

[tool.setuptools]
packages = [
    "mkdocs_image_gallery_plugin",
    "mkdocs_image_gallery_plugin.assets",
    "mkdocs_image_gallery_plugin.assets.css",
    "mkdocs_image_gallery_plugin.assets.js",
    "mkdocs_image_gallery_plugin.assets.templates",
]
include-package-data = true